requirements.txt
streamlit
transformers
torch
Pillow
pdfminer.six 
sentence-transformers
langchain
chromadb
streamlit_chat
accelerate
